Respond to code review feedback from my team.
* The WireMock http port was not being assigned using the dynamic
assignment feature that was built into the test because it defaulted to
8080 in the static initializer.
* There was an unnecessary and somewhat error prone check around the
unsupported operation exception in resolveByURL that was copied from the
HAPI JPA implementation
* The Bundle.isEmpty() method does not behave with the same semantics as
checking to Bundle.getEntry().isEmpty() and is potentially error prone.
* The display value in lookup would potentially overwrite a provided
display value with null when the parameter was not provided in the
lookup response. Defaulting to no touch now when it isn't there.