Update resolveByUrl search path, switch expand to GET
Per feedback, I updated the resolveByUrl method to use a search path
that looks first for Resource.url, then Resource.identifier, and, if all
else fails *and* the value looks like a FHIR resource ID (stripping
urn:oid: or urn:uuid: prefixes) it tries to read by Resource.id.

This changeset also switches the expand operation to useHttpGet() vs.
the default POST operation because there are problems with POST'ing an
empty parameters object.

Please excuse the whitespace changes in the test class. My editor was
using tabs and the project seems to prefer spaces, so I updated the file
to be consistent with the rest of the project.