#463 - Add ability to configure search page size
The default _count value for FHIR servers can be quite low (10-20
records at a time) resulting in a large number of REST calls to the
server when retrieving data necessary for CQL evaluation. This parameter
enables the user to configure how many rows should be returned per
request to the server.