#469 Fix broken SLF4J logging in tests
The slf4j-simple dependency got auto-upgraded as part of the maintenance
script execution. This created a problem where the API and log bindings
were on different versions and the log binding was not getting loaded
property. I updated the POM to explicitly declare the API and log
binding versions and use a property to make sure they are in sync. I
also excluded these from getting automatically upgraded to the 2.x alpha
line of releases next time the maintenance script runs.